backend/requirements.txt
flask
flask-cors
gunicorn
flask-mail
pymongo[srv]
dnspython
python-dotenv
google-genai
